Add regression pipeline (#885)
* Add regression pipeline that build with itwinjs-core ^4.10.0

* Add build against 4.0.0

* Override devDependencies instead of all

* Update lock file

* Log all failures

* Build with 4.4.0

* Run tests

* Fix full-stack-tests-build

* Remove 4.4.0 build

* Handle older versions in full stack tests

* Update script

* Path build errors

* Update favorite properties tests

* Run regression test using locally built tarballs

* No frozen lockfile

* Update paths

* Correct folder names

* Add some logging

* Flatten tarballs directory

* Remove old regression pipeline

* Apply patches based on version

* Apply suggestions from code review


* Apply suggestions from code review


* Address PR comments

* Update diff

* Build with different AppUI versions

* Prettier

* Fix lock-fie

* Fix import

* Prettier...

* Replace `assert` -> `with`

* Add lint ignore

* Parse dependencies from workspace file

---------

+++ apps/full-stack-tests/src/unified-selection/learning-snippets/ReadmeExample.test.tsx
@@ -24,7 +24,7 @@ import { Presentation } from "@itwin/presentation-frontend";
 import { buildIModel } from "../../IModelUtils.js";
 import { initialize, terminate } from "../../IntegrationTests.js";
 import { render, waitFor } from "../../RenderUtils.js";
-import { stubGetBoundingClientRect } from "../../Utils.js";
+import { isSelectionStorageSupported, stubGetBoundingClientRect } from "../../Utils.js";
 
 describe("Unified selection", () => {
   describe("Learning snippets", () => {
@@ -125,35 +125,37 @@ describe("Unified selection", () => {
         await waitFor(() => expect(getByText("Number of selected elements: 1")).to.not.be.null);
       });
 
-      it("Unified selection sync with legacy SelectionManager", async function () {
-        Presentation.terminate();
+      if (isSelectionStorageSupported()) {
+        it("Unified selection sync with legacy SelectionManager", async function () {
+          Presentation.terminate();
 
-        const { imodel, ...keys } = await buildIModel(this, async (builder) => {
-          const modelKey = insertPhysicalModelWithPartition({ builder, codeValue: "test model" });
-          const categoryKey = insertSpatialCategory({ builder, codeValue: "test category" });
-          const elementKey = insertPhysicalElement({ builder, userLabel: "root element", modelId: modelKey.id, categoryId: categoryKey.id });
-          return { modelKey, categoryKey, elementKey };
+          const { imodel, ...keys } = await buildIModel(this, async (builder) => {
+            const modelKey = insertPhysicalModelWithPartition({ builder, codeValue: "test model" });
+            const categoryKey = insertSpatialCategory({ builder, codeValue: "test category" });
+            const elementKey = insertPhysicalElement({ builder, userLabel: "root element", modelId: modelKey.id, categoryId: categoryKey.id });
+            return { modelKey, categoryKey, elementKey };
+          });
+
+          // __PUBLISH_EXTRACT_START__ Presentation.LegacySelectionManagerSelectionSync.Example
+          const selectionStorage = createStorage();
+
+          // Initialize Presentation with our selection storage, to make sure that any components, using `Presentation.selection`,
+          // use the same underlying selection store.
+          await Presentation.initialize({
+            selection: {
+              selectionStorage,
+            },
+          });
+          // __PUBLISH_EXTRACT_END__
+
+          expect(Selectables.isEmpty(selectionStorage.getSelection({ imodelKey: imodel.key }))).to.be.true;
+
+          Presentation.selection.addToSelection("test", imodel, new KeySet([keys.elementKey]));
+          await waitFor(() => {
+            expect(Selectables.size(selectionStorage.getSelection({ imodelKey: imodel.key }))).to.eq(1);
+          });
         });
-
-        // __PUBLISH_EXTRACT_START__ Presentation.LegacySelectionManagerSelectionSync.Example
-        const selectionStorage = createStorage();
-
-        // Initialize Presentation with our selection storage, to make sure that any components, using `Presentation.selection`,
-        // use the same underlying selection store.
-        await Presentation.initialize({
-          selection: {
-            selectionStorage,
-          },
-        });
-        // __PUBLISH_EXTRACT_END__
-
-        expect(Selectables.isEmpty(selectionStorage.getSelection({ imodelKey: imodel.key }))).to.be.true;
-
-        Presentation.selection.addToSelection("test", imodel, new KeySet([keys.elementKey]));
-        await waitFor(() => {
-          expect(Selectables.size(selectionStorage.getSelection({ imodelKey: imodel.key }))).to.eq(1);
-        });
-      });
+      }
     });
   });
 });
